+++ src/visaplan/plone/staticthumbnails/interfaces.py
@@ -1,8 +1,12 @@
 # -*- coding: utf-8 -*-
 """Module where all interfaces, events and exceptions live."""
 
-from zope.publisher.interfaces.browser import IDefaultBrowserLayer
+# Python compatibility:
+from __future__ import absolute_import
+
+# Zope:
 from zope.interface import Interface
+from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 
 
 class IVisaplanPloneStaticthumbnailsLayer(IDefaultBrowserLayer):