+++ src/visaplan/plone/staticthumbnails/setuphandlers.py
@@ -1,4 +1,8 @@
 # -*- coding: utf-8 -*-
+# Python compatibility:
+from __future__ import absolute_import
+
+# Zope:
 from Products.CMFPlone.interfaces import INonInstallable
 from zope.interface import implementer