+++ src/visaplan/plone/staticthumbnails/testing.py
@@ -1,12 +1,22 @@
 # -*- coding: utf-8 -*-
-from plone.app.robotframework.testing import REMOTE_LIBRARY_BUNDLE_FIXTURE
-from plone.app.testing import applyProfile
-from plone.app.testing import FunctionalTesting
-from plone.app.testing import IntegrationTesting
-from plone.app.testing import PLONE_FIXTURE
-from plone.app.testing import PloneSandboxLayer
-from plone.testing import z2
+# Python compatibility:
+from __future__ import absolute_import
 
+# Standard library:
+import warnings
+
+with warnings.catch_warnings():
+    warnings.simplefilter('ignore', ImportWarning)
+
+    # from plone.app.robotframework.testing import REMOTE_LIBRARY_BUNDLE_FIXTURE
+    from plone.app.testing import applyProfile
+    from plone.app.testing import FunctionalTesting
+    from plone.app.testing import IntegrationTesting
+    from plone.app.testing import PLONE_FIXTURE
+    from plone.app.testing import PloneSandboxLayer
+    from plone.testing import z2
+
+# visaplan:
 import visaplan.plone.staticthumbnails
 
 
@@ -18,6 +28,7 @@ class VisaplanPloneStaticthumbnailsLayer(PloneSandboxLayer):
         # Load any other ZCML that is required for your tests.
         # The z3c.autoinclude feature is disabled in the Plone fixture base
         # layer.
+        # Plone:
         import plone.app.dexterity
         self.loadZCML(package=plone.app.dexterity)
         self.loadZCML(package=visaplan.plone.staticthumbnails)
@@ -41,11 +52,11 @@ VISAPLAN_PLONE_STATICTHUMBNAILS_FUNCTIONAL_TESTING = FunctionalTesting(
 )
 
 
-VISAPLAN_PLONE_STATICTHUMBNAILS_ACCEPTANCE_TESTING = FunctionalTesting(
-    bases=(
-        VISAPLAN_PLONE_STATICTHUMBNAILS_FIXTURE,
-        REMOTE_LIBRARY_BUNDLE_FIXTURE,
-        z2.ZSERVER_FIXTURE,
-    ),
-    name='VisaplanPloneStaticthumbnailsLayer:AcceptanceTesting',
-)
+# VISAPLAN_PLONE_STATICTHUMBNAILS_ACCEPTANCE_TESTING = FunctionalTesting(
+#     bases=(
+#         VISAPLAN_PLONE_STATICTHUMBNAILS_FIXTURE,
+#         REMOTE_LIBRARY_BUNDLE_FIXTURE,
+#         z2.ZSERVER_FIXTURE,
+#     ),
+#     name='VisaplanPloneStaticthumbnailsLayer:AcceptanceTesting',
+# )