+++ CHANGES.rst
@@ -3,8 +3,20 @@ Changelog
 =========
 
 
-1.0.dev1 (unreleased)
----------------------
+1.0.1 (unreleased)
+------------------
+
+- Python 3 compatibility, using six_
+- imports sorted by isort_
+
+[tobiasherp]
+
+
+1.0 (2019-06-26)
+----------------
 
 - Initial release.
   [tobiasherp]
+
+.. _isort: https://pypi.org/project/isort
+.. _six: https://pypi.org/project/six